Remove imports of taurus.core.tango in taurus.qt submodule
Remove module-level imports of taurus.core.tango in taurus.qt submodule.
In some cases, use taurus-equivalent code and in others delay the
import and mark it as tango-centric.
