+++ lib/taurus/qt/qtgui/panel/taurusmodelchooser.py
@@ -37,9 +37,6 @@ from taurus.qt.qtgui.tree import TaurusDbTreeWidget
 from taurus.core.util.containers import CaselessList
 from taurusmodellist import TaurusModelList
 
-#@todo: tango-centric!!
-from taurus.core.tango.tangodatabase import TangoDevInfo, TangoAttrInfo
-
 
 class TaurusModelSelectorTree(TaurusWidget):
 
@@ -108,6 +105,8 @@ class TaurusModelSelectorTree(TaurusWidget):
 
     def getSelectedModels(self):
         # todo: this method is tango-centric, but it could be fixed...
+        # @todo: tango-centric!!
+        from taurus.core.tango.tangodatabase import TangoDevInfo, TangoAttrInfo
         selected = []
         for item in self._deviceTree.selectedItems():
             nfo = item.itemData()
