+++ lib/taurus/qt/qtgui/tree/taurusdevicetree.py
@@ -37,6 +37,7 @@ __all__ = ["TaurusDevTree", "TaurusSearchTree", "TaurusDevTreeOptions"]
 
 import time
 import os
+import re
 import traceback
 from functools import partial
 
@@ -50,12 +51,19 @@ except:
 
 from taurus.external.qt import Qt
 
+import taurus
 import taurus.core
 from taurus.core.util.colors import DEVICE_STATE_PALETTE, ATTRIBUTE_QUALITY_PALETTE
 from taurus.core.util.containers import CaselessDict
-from taurus.core.tango.search import *  # @TODO: Avoid implicit imports
+from taurus.core.util.fandango_search import (
+    isCallable, isString, split_model_list, isSequence, isMap,
+    get_matching_devices, matchCl, get_alias_for_device, extend_regexp)
 from taurus.qt.qtcore.util.emitter import SingletonWorker
-from taurus.qt.qtcore.mimetypes import *  # @TODO: Avoid implicit imports
+from taurus.qt.qtcore.mimetypes import (TAURUS_MODEL_LIST_MIME_TYPE,
+                                        TAURUS_DEV_MIME_TYPE,
+                                        TAURUS_ATTR_MIME_TYPE,
+                                        TAURUS_MODEL_MIME_TYPE
+                                        )
 from taurus.qt.qtcore.util import properties
 from taurus.qt.qtcore.util.properties import djoin
 from taurus.qt.qtgui.base import TaurusBaseComponent, TaurusBaseWidget
@@ -211,6 +219,7 @@ class TaurusTreeNodeContainer(object):
 
     def getNodeDraggable(self, node=None):
         """ This method will return True only if the selected node belongs to a numeric Tango attribute """
+        import PyTango  # TODO: tango-centric
         numtypes = [PyTango.DevDouble, PyTango.DevFloat, PyTango.DevLong, PyTango.DevLong64,
                     PyTango.DevULong, PyTango.DevShort, PyTango.DevUShort, PyTango.DevBoolean]
         if node is None:
